consumer/requirements.txt
psycopg2-binary==2.9.4
kafka-python==2.0.2
